Fix problem with symlink creation for windows hosts / vagrant
Appears to be a known issue with symbolic links when a Windows
host is running our DevBox guest trying to create symbolic links
in a vagrant shared folder:

https://www.virtualbox.org/ticket/10085

For now to fix/work around we just copy the files instead of
symlinking.  That link gives a possible fix to allow
symlinks on windows hosts, which may be better, but will need
to test it a bit first.