Add const overloaded operator[]
Add a const member function of overloaded operator[] to the
List class.  This overloaded operator[] returns a copy of
the value at the indicated index, instead of a reference
to the value.  This method allows a const List& (a constant
reference) to be passed in as a parameter for the merge
and other functions in this assignment, since this new
version of the operator can be used to read out values
from the List.